pyproject.toml
[project]
name = "ufc-pokedex"
version = "0.1.0"
description = "UFC fighter statistics and information database"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.110.0",
    "uvicorn[standard]==0.29.0",
    "sqlalchemy==2.0.23",
    "greenlet==3.0.3",
    "alembic==1.13.1",
    "psycopg[binary]==3.2.12",
    "aiosqlite>=0.19.0",
    "pydantic==2.7.1",
    "pydantic-settings==2.2.1",
    "scrapy==2.11.1",
    "beautifulsoup4==4.12.3",
    "httpx==0.27.0",
    "lxml==5.2.1",
    "tenacity==8.2.3",
    "python-dotenv==1.0.1",
    "rich==13.7.1",
    "rapidfuzz==3.10.1",
    "redis==5.0.4",
    "pandas>=2.0.0",
    "unidecode>=1.3.0",
    "requests>=2.31.0",
    "opencv-python>=4.8.1",
    "dlib>=19.24.2",
    "numpy>=1.26.0",
    "pillow>=10.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.1.1",
    "pytest-asyncio==0.23.5",
    "pytest-cov==4.1.0",
    "responses==0.25.0",
    "ruff==0.3.5",
    "mypy==1.8.0",
    "types-beautifulsoup4==4.12.0.7",
    "types-requests==2.31.0.20240311",
    "pre-commit==3.6.2",
]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["frontend"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "S", "C4"]
ignore = ["B008"]  # Allow function default arguments with calls (use carefully)

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q --disable-warnings"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
ignore_missing_imports = true

[tool.coverage.run]
source = ["scraper", "backend"]
branch = true
